+++ src/main/java/onlyoffice/OnlyOfficeEditorServlet.java
@@ -32,6 +32,8 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.atlassian.plugin.webresource.WebResourceUrlProvider;
+import com.atlassian.plugin.webresource.UrlMode;
 import onlyoffice.managers.configuration.ConfigurationManager;
 import onlyoffice.managers.convert.ConvertManager;
 import onlyoffice.managers.document.DocumentManager;
@@ -66,6 +68,8 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
     private final SettingsManager settingsManager;
     @ComponentImport
     private final LocaleManager localeManager;
+    @ComponentImport
+    private final WebResourceUrlProvider webResourceUrlProvider;
 
     private final JwtManager jwtManager;
     private final UrlManager urlManager;
@@ -78,10 +82,9 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
 
     @Inject
     public OnlyOfficeEditorServlet(PluginSettingsFactory pluginSettingsFactory, LocaleManager localeManager,
-                                   SettingsManager settingsManager, UrlManager urlManager,
-                                   JwtManager jwtManager, ConfigurationManager configurationManager,
-                                   AuthContext authContext, DocumentManager documentManager,
-                                   AttachmentUtil attachmentUtil, ConvertManager convertManager) {
+            SettingsManager settingsManager, UrlManager urlManager, JwtManager jwtManager,
+            ConfigurationManager configurationManager, AuthContext authContext, DocumentManager documentManager,
+            AttachmentUtil attachmentUtil, ConvertManager convertManager, WebResourceUrlProvider webResourceUrlProvider) {
         this.pluginSettingsFactory = pluginSettingsFactory;
         this.localeManager = localeManager;
         this.settingsManager = settingsManager;
@@ -92,6 +95,7 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
         this.documentManager = documentManager;
         this.attachmentUtil = attachmentUtil;
         this.convertManager = convertManager;
+        this.webResourceUrlProvider = webResourceUrlProvider;
     }
 
     @Override
@@ -185,10 +189,16 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
         String docTitle = fileName.trim();
         String docExt = docTitle.substring(docTitle.lastIndexOf(".") + 1).trim().toLowerCase();
         boolean canEdit = documentManager.isEditable(docExt);
+        String documentType = documentManager.getDocType(docExt);
 
         config.put("docserviceApiUrl", apiUrl + properties.getProperty("files.docservice.url.api"));
         config.put("errorMessage", errorMessage);
         config.put("docTitle", docTitle);
+        config.put("favicon", webResourceUrlProvider.getStaticPluginResourceUrl(
+                "onlyoffice.onlyoffice-confluence-plugin:onlyoffice-confluence-plugin-resources-editor",
+                documentType +".ico",
+                UrlMode.ABSOLUTE)
+        );
 
         JSONObject responseJson = new JSONObject();
         JSONObject documentObject = new JSONObject();
@@ -202,7 +212,7 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
             responseJson.put("type", "desktop");
             responseJson.put("width", "100%");
             responseJson.put("height", "100%");
-            responseJson.put("documentType", documentManager.getDocType(docExt));
+            responseJson.put("documentType", documentType);
 
             responseJson.put("document", documentObject);
             documentObject.put("title", docTitle);