+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - open non-editable formats on views
 - added formats for conversion docm, dot, dotx, epub, htm, html, otp, ots, ott, pot, potm, potx, pps, ppsm, ppsx,
  pptm, rtf, xlsm, xlt, xltm, xltx
+- change favicon in editor by document type
 
 ## 3.0.0
 ## Added